Added linters
Also fixed existing linting issues and a few syntantical errors..
Added a setup.py/cfg file
