+++ setup.cfg
@@ -1,0 +1,46 @@
+[metadata]
+name = cerberus
+description = TO guard cluster
+author = chaitanyaenr
+author-email = nelluri@redhat.com
+license = mit
+long-description = file: README.md
+long-description-content-type = text/markdown; charset=UTF-8
+classifiers =
+    Development Status :: 4 - Beta
+    Programming Language :: Python
+
+[options]
+zip_safe = False
+packages = find:
+include_package_data = True
+package_dir =
+    =src
+# Add here dependencies of your project (semicolon/line-separated), e.g.
+install_requires = PyYAML
+# tests_require = pytest; pytest-cov
+# Require a specific Python version, e.g. Python 2.7 or >= 3.4
+# python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*
+
+[options.packages.find]
+where = src
+
+[options.extras_require]
+# Add here additional requirements for extra features, to install with:
+# `pip install touchstone[PDF]` like:
+# PDF = ReportLab; RXP
+
+[aliases]
+dists = bdist_wheel
+
+[bdist_wheel]
+# Use this option if your package is pure-python
+universal = 1
+
+[flake8]
+# Some sane defaults for the code style checker flake8
+exclude =
+    .tox
+    build
+    dist
+    .eggsse
