+++ tox.ini
@@ -1,0 +1,23 @@
+[tox]
+envlist = pep8
+
+[testenv]
+usedevelop = True
+setenv =
+   VIRTUAL_ENV={envdir}
+deps = -r{toxinidir}/test-requirements.txt
+commands = python setup.py develop
+
+[testenv:pep8]
+basepython = python
+commands = flake8 {posargs}
+
+[testenv:venv]
+commands = {posargs}
+
+[flake8]
+# E123, E125 skipped as they are invalid PEP-8.
+show-source = True
+ignore = E123,E125
+builtins = _
+exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build
